+++ src/main/java/roomescape/presentation/web/controller/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.presentation.web.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
