+++ src/main/java/roomescape/presentation/web/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.presentation.web.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.request.CreateReservationRequest;
-import roomescape.controller.response.ReservationResponse;
-import roomescape.controller.response.ReservationTimeResponse;
 import roomescape.domain.Name;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
+import roomescape.presentation.web.request.CreateReservationRequest;
+import roomescape.presentation.web.response.ReservationResponse;
+import roomescape.presentation.web.response.ReservationTimeResponse;
 
 @RestController
 @RequestMapping("/reservations")
