+++ src/main/java/roomescape/presentation/web/request/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.request;
+package roomescape.presentation.web.request;
 
 import java.time.LocalDate;
