+++ src/main/java/roomescape/presentation/web/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.response;
+package roomescape.presentation.web.response;
 
 import java.time.LocalTime;
 import com.fasterxml.jackson.annotation.JsonFormat;
